Register anchors in ServerRedoc component
This uses the `useBrokenLinks` Docusaurus hook to register Redoc-generated
anchors so that Redocusaurus users can configure their sites to fail to
build if there are broken anchors.

The ServerRedoc component is the only one that registers anchors because
if a Redoc page is rendered on the client side it is too late to register
anchors for the Docusaurus build.